# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GAMEA Social Monitor",
    "description": "API Institucional de Monitoreo y Analítica de Interacciones en Redes Sociales del Gobierno Autónomo Municipal de El Alto (GAMEA).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/publications/": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "List Publications",
        "description": "Consulta paginada de publicaciones institucionales sujetas a monitoreo con métricas calculadas.",
        "operationId": "list_publications_api_v1_publications__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "platform_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por plataforma",
              "title": "Platform Id"
            },
            "description": "Filtrar por plataforma"
          },
          {
            "name": "campaign_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por campaña asociada",
              "title": "Campaign Id"
            },
            "description": "Filtrar por campaña asociada"
          },
          {
            "name": "is_monitored",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por estado de monitoreo",
              "title": "Is Monitored"
            },
            "description": "Filtrar por estado de monitoreo"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Número de página (1-indexado)",
              "default": 1,
              "title": "Page"
            },
            "description": "Número de página (1-indexado)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Registros por página (máximo 1000)",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Registros por página (máximo 1000)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_PublicationResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Publications"
        ],
        "summary": "Create Publication",
        "description": "Registra una publicación institucional para monitoreo (REQ-PUB-002).\nOperación idempotente (Principio XI): Si el post_id ya existe, retorna el registro existente.",
        "operationId": "create_publication_api_v1_publications__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publications/facebook/page-posts": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "Get Facebook Recent Posts",
        "description": "Retorna las publicaciones más recientes de la página oficial de Facebook GAMEA para importación rápida.",
        "operationId": "get_facebook_recent_posts_api_v1_publications_facebook_page_posts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FacebookRecentPostItem"
                  },
                  "type": "array",
                  "title": "Response Get Facebook Recent Posts Api V1 Publications Facebook Page Posts Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/publications/import-from-url": {
      "post": {
        "tags": [
          "Publications"
        ],
        "summary": "Import Publication From Url",
        "description": "Extrae el ID y registra automáticamente un post desde un link de Facebook o TikTok para monitoreo.",
        "operationId": "import_publication_from_url_api_v1_publications_import_from_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPostFromUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/publications/{publication_id}": {
      "get": {
        "tags": [
          "Publications"
        ],
        "summary": "Get Publication",
        "description": "Consulta el detalle de una publicación monitoreada.",
        "operationId": "get_publication_api_v1_publications__publication_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Publication Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/campaigns/{campaign_id}/publications": {
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Add Publications To Campaign",
        "description": "Vincula publicaciones a una campaña de monitoreo.",
        "operationId": "add_publications_to_campaign_api_v1_campaigns__campaign_id__publications_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPublicationsToCampaignRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringCampaignDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddPublicationsToCampaignRequest": {
        "properties": {
          "publication_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Publication Ids",
            "description": "Lista de IDs de publicaciones a vincular"
          }
        },
        "type": "object",
        "required": [
          "publication_ids"
        ],
        "title": "AddPublicationsToCampaignRequest"
      },
...
      "FacebookRecentPostItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "created_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Time"
          },
          "permalink_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permalink Url"
          },
          "shares_count": {
            "type": "integer",
            "title": "Shares Count",
            "default": 0
          },
          "is_monitored": {
            "type": "boolean",
            "title": "Is Monitored",
            "default": false
          },
          "existing_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Existing Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "FacebookRecentPostItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportPostFromUrlRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Enlace completo de la publicación en Facebook o TikTok"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Título o descripción personalizada opcional"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "description": "FACEBOOK o TIKTOK",
            "default": "FACEBOOK"
          },
          "campaign_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Id"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ImportPostFromUrlRequest"
      },
...
      "MonitoringCampaignDetailResponse": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title",
            "description": "Título de la campaña"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "publications": {
            "items": {
              "$ref": "#/components/schemas/PublicationResponse"
            },
            "type": "array",
            "title": "Publications",
            "default": []
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/MonitoringTargetResponse"
            },
            "type": "array",
            "title": "Targets",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "MonitoringCampaignDetailResponse"
      },
...
      "PageResponse_PublicationResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicationResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "Lista de elementos de la página actual"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total general de elementos"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "Número de página actual"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size",
            "description": "Tamaño de la página"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "description": "Total de páginas disponibles"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "PageResponse[PublicationResponse]"
      },
...
      "PublicationCreate": {
        "properties": {
          "platform_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "title": "Platform Id"
          },
          "institutional_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutional Account Id"
          },
          "external_post_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "External Post Id",
            "description": "ID del post en Facebook/TikTok"
          },
          "post_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Url"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "default": "POST"
          },
          "is_monitored": {
            "type": "boolean",
            "title": "Is Monitored",
            "default": true
          },
          "campaign_ids": {
            "items": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "array",
            "title": "Campaign Ids",
            "description": "Campañas a las que se asocia",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "platform_id",
          "external_post_id"
        ],
        "title": "PublicationCreate"
      },
      "PublicationResponse": {
        "properties": {
          "platform_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "string"
              }
            ],
            "title": "Platform Id"
          },
          "institutional_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institutional Account Id"
          },
          "external_post_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "External Post Id",
            "description": "ID del post en Facebook/TikTok"
          },
          "post_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Url"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "content_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Text"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type",
            "default": "POST"
          },
          "is_monitored": {
            "type": "boolean",
            "title": "Is Monitored",
            "default": true
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "platform_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Name"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "platform": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SocialPlatformResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_reactions": {
            "type": "integer",
            "title": "Total Reactions",
            "default": 0
          },
          "total_comments": {
            "type": "integer",
            "title": "Total Comments",
            "default": 0
          },
          "total_shares": {
            "type": "integer",
            "title": "Total Shares",
            "default": 0
          },
          "meta_reactions_total": {
            "type": "integer",
            "title": "Meta Reactions Total",
            "description": "Total de reacciones reportado por Meta Graph API (conteo agregado, sin identidades)",
            "default": 0
          },
          "meta_reactions_by_type": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Meta Reactions By Type",
            "description": "Desglose por tipo (like, love, wow, haha, sad, angry) reportado por Meta"
          },
          "meta_metrics_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Metrics Synced At",
            "description": "UTC de la última lectura de métricas agregadas de Meta"
          }
        },
        "type": "object",
        "required": [
          "platform_id",
          "external_post_id",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PublicationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}